+++ package.json
@@ -40,7 +40,8 @@
     "mongoose-unique-validator": "^2.0.2",
     "morgan": "^1.9.1",
     "node-schedule": "^1.3.0",
-    "passport": "^0.4.0"
+    "passport": "^0.4.0",
+    "recipe-ingredient-parser": "^1.5.6"
   },
   "devDependencies": {
     "eslint": "^5.13.0",
