+++ views/partials/navbar.hbs
@@ -21,8 +21,11 @@
       <div class="nav-links">
         <!-- show these links if the user is logged in  -->
         {{#if user}}
-        <li class="nav-item">
+        {{!-- <li class="nav-item">
           <a class="nav-link nav-dash" href="/dashboard">dashboard</a>
+        </li> --}}
+        <li class="nav-item">
+          <a class="nav-link nav-favs" href="/dashboard/favorites">favorites</a>
         </li>
         <li class="nav-item">
           <a href="/cart" class="nav-link nav-cart">cart</a>
